+++ packages/amplify-category-api/src/__tests__/provider-utils/awscloudformation/utils/rds-utils.test.ts
@@ -1,22 +1,54 @@
-import { DataSourceStrategiesProvider, ModelDataSourceStrategy } from '@aws-amplify/graphql-transformer-interfaces';
+import {
+  DataSourceStrategiesProvider,
+  ModelDataSourceStrategy,
+  SQLLambdaModelDataSourceStrategy,
+} from '@aws-amplify/graphql-transformer-interfaces';
 import { DDB_DEFAULT_DATASOURCE_STRATEGY, MYSQL_DB_TYPE } from '@aws-amplify/graphql-transformer-core';
-import { checkForUnsupportedDirectives } from '../../../../provider-utils/awscloudformation/utils/rds-resources/utils';
+import {
+  checkForUnsupportedDirectives,
+  containsSqlModelOrDirective,
+} from '../../../../provider-utils/awscloudformation/utils/rds-resources/utils';
 
 describe('check for unsupported RDS directives', () => {
+  const dbConnectionConfig = {
+    databaseNameSsmPath: '/databaseNameSsmPath',
+    hostnameSsmPath: '/hostnameSsmPath',
+    portSsmPath: '/portSsmPath',
+    usernameSsmPath: '/usernameSsmPath',
+    passwordSsmPath: '/passwordSsmPath',
+  };
+
   const dataSourceStrategies: Record<string, ModelDataSourceStrategy> = {
     Post: {
       name: 'mysqlstrategy',
       dbType: MYSQL_DB_TYPE,
-      dbConnectionConfig: {
-        databaseNameSsmPath: '/databaseNameSsmPath',
-        hostnameSsmPath: '/hostnameSsmPath',
-        portSsmPath: '/portSsmPath',
-        usernameSsmPath: '/usernameSsmPath',
-        passwordSsmPath: '/passwordSsmPath',
-      },
+      dbConnectionConfig,
     },
     Tag: DDB_DEFAULT_DATASOURCE_STRATEGY,
   };
+
+  const ddbDataSourceStrategies: Record<string, ModelDataSourceStrategy> = {
+    Post: DDB_DEFAULT_DATASOURCE_STRATEGY,
+    Tag: DDB_DEFAULT_DATASOURCE_STRATEGY,
+  };
+
+  const strategy: SQLLambdaModelDataSourceStrategy = {
+    name: 'strategy',
+    dbType: 'MYSQL',
+    dbConnectionConfig,
+  };
+
+  const sqlDirectiveDataSourceStrategies = [
+    {
+      typeName: 'Query' as const,
+      fieldName: 'myCustomQuery',
+      strategy: strategy,
+      customSqlStatements: {
+        myCustomQuery: 'SELECT 1;',
+        myCustomMutation: 'UPDATE mytable SET id=1; SELECT 1;',
+      },
+    },
+  ];
 
   const dataSourceStrategiesProvider: DataSourceStrategiesProvider = { dataSourceStrategies };
   const emptyProvider: DataSourceStrategiesProvider = { dataSourceStrategies: {} };
@@ -143,5 +175,18 @@ describe('check for unsupported RDS directives', () => {
   it('early return if schema is empty or undefined', () => {
     const schema = '';
     expect(() => checkForUnsupportedDirectives(schema, dataSourceStrategiesProvider)).not.toThrowError();
+  });
+
+  it('containsSqlModelOrDirective should return true if there are sql models', () => {
+    expect(containsSqlModelOrDirective(dataSourceStrategies, undefined)).toBeTruthy();
+    expect(containsSqlModelOrDirective(dataSourceStrategies, [])).toBeTruthy();
+    expect(containsSqlModelOrDirective(emptyProvider.dataSourceStrategies, sqlDirectiveDataSourceStrategies)).toBeTruthy();
+    expect(containsSqlModelOrDirective(ddbDataSourceStrategies, sqlDirectiveDataSourceStrategies)).toBeTruthy();
+  });
+
+  it('containsSqlModelOrDirective should return false if there are no sql models', () => {
+    expect(containsSqlModelOrDirective(ddbDataSourceStrategies, undefined)).toBeFalsy();
+    expect(containsSqlModelOrDirective(ddbDataSourceStrategies, [])).toBeFalsy();
+    expect(containsSqlModelOrDirective(emptyProvider.dataSourceStrategies)).toBeFalsy();
   });
 });