+++ packages/amplify-graphql-api-construct/src/internal/codegen-assets.ts
@@ -1,5 +1,5 @@
 import { RemovalPolicy } from 'aws-cdk-lib';
-import { Bucket } from 'aws-cdk-lib/aws-s3';
+import { Bucket, IBucket } from 'aws-cdk-lib/aws-s3';
 import { BucketDeployment, Source } from 'aws-cdk-lib/aws-s3-deployment';
 import { Construct } from 'constructs';
 
@@ -15,7 +15,7 @@ const MODEL_SCHEMA_KEY = 'model-schema.graphql';
  * @param key the key to embed in the uri string
  * @returns the uri string representation.
  */
-const getS3UriForBucketAndKey = (bucket: Bucket, key: string): string => `s3://${bucket.bucketName}/${key}`;
+const getS3UriForBucketAndKey = (bucket: IBucket, key: string): string => `s3://${bucket.bucketName}/${key}`;
 
 /**
  * Construct which creates a bucket, and uploads file assets required for codegen to run.
@@ -32,7 +32,7 @@ export class CodegenAssets extends Construct {
       autoDeleteObjects: true,
     });
 
-    new BucketDeployment(this, `${id}Deployment`, {
+    const deployment = new BucketDeployment(this, `${id}Deployment`, {
       destinationBucket: bucket,
       sources: [Source.data(MODEL_SCHEMA_KEY, props.modelSchema)],
       // Bucket deployment uses a Lambda that runs AWS S3 CLI to transfer assets to destination bucket.
@@ -42,6 +42,6 @@ export class CodegenAssets extends Construct {
       memoryLimit: 1536,
     });
 
-    this.modelSchemaS3Uri = getS3UriForBucketAndKey(bucket, MODEL_SCHEMA_KEY);
+    this.modelSchemaS3Uri = getS3UriForBucketAndKey(deployment.deployedBucket, MODEL_SCHEMA_KEY);
   }
 }