+++ packages/amplify-graphql-auth-transformer/src/utils/definitions.ts
@@ -32,6 +32,7 @@ export interface SearchableConfig {
 
 export type GetAuthRulesOptions = GetArgumentsOptions & {
   isField?: boolean;
+  isSqlDataSource?: boolean;
 };
 
 /**