+++ packages/amplify-graphql-transformer-migrator/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.17](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-migrator@2.2.16...@aws-amplify/graphql-transformer-migrator@2.2.17) (2024-01-30)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-migrator
+
+## [2.2.16](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-migrator@2.2.15...@aws-amplify/graphql-transformer-migrator@2.2.16) (2024-01-22)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-migrator
+
 ## [2.2.15](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-migrator@2.2.14...@aws-amplify/graphql-transformer-migrator@2.2.15) (2023-12-18)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer-migrator